backend/requirements.txt
boto3==1.34.34
email-validator==2.1.0
bleach==6.1.0
python-dotenv==1.0.0
fastapi==0.109.0
mangum==0.17.0
pydantic>=2.10.0
uvicorn[standard]>=0.27.0
